+++ hsdp/provider.go
@@ -5,6 +5,7 @@ import (
 	"github.com/google/fhir/go/jsonformat"
 	"github.com/hashicorp/terraform-plugin-sdk/v2/diag"
 	"github.com/hashicorp/terraform-plugin-sdk/v2/helper/schema"
+	"os"
 )
 
 // Provider returns an instance of the HSDP provider
@@ -12,15 +13,15 @@ func Provider(build string) *schema.Provider {
 	return &schema.Provider{
 		Schema: map[string]*schema.Schema{
 			"region": {
-				Type:         schema.TypeString,
-				Optional:     true,
-				RequiredWith: []string{"environment"},
-				Description:  descriptions["region"],
-			},
-			"environment": {
 				Type:        schema.TypeString,
 				Optional:    true,
-				Description: descriptions["environment"],
+				Description: descriptions["region"],
+			},
+			"environment": {
+				Type:         schema.TypeString,
+				Optional:     true,
+				Description:  descriptions["environment"],
+				RequiredWith: []string{"region"},
 			},
 			"iam_url": {
 				Type:        schema.TypeString,
@@ -32,10 +33,10 @@ func Provider(build string) *schema.Provider {
 				Optional:    true,
 				Description: descriptions["idm_url"],
 			},
-			"credentials_url": {
+			"s3creds_url": {
 				Type:        schema.TypeString,
 				Optional:    true,
-				Description: descriptions["credentials_url"],
+				Description: descriptions["s3creds_url"],
 			},
 			"service_id": {
 				Type:          schema.TypeString,
@@ -174,11 +175,21 @@ func Provider(build string) *schema.Provider {
 			"hsdp_iam_mfa_policy":      resourceIAMMFAPolicy(),
 			"hsdp_iam_password_policy": resourceIAMPasswordPolicy(),
 			"hsdp_iam_email_template":  resourceIAMEmailTemplate(),
-			"hsdp_credentials_policy":  resourceCredentialsPolicy(),
+			"hsdp_s3creds_policy":      resourceS3CredsPolicy(),
 			"hsdp_container_host":      resourceContainerHost(),
+			"hsdp_container_host_exec": resourceContainerHostExec(),
 			"hsdp_metrics_autoscaler":  resourceMetricsAutoscaler(),
 			"hsdp_cdr_org":             resourceCDROrg(),
 			"hsdp_cdr_subscription":    resourceCDRSubscription(),
+			"hsdp_dicom_store_config":  resourceDICOMStoreConfig(),
+			"hsdp_dicom_object_store":  resourceDICOMObjectStore(),
+			"hsdp_dicom_repository":    resourceDICOMRepository(),
+			"hsdp_pki_tenant":          resourcePKITenant(),
+			"hsdp_pki_cert":            resourcePKICert(),
+			"hsdp_stl_app":             resourceSTLApp(),
+			"hsdp_stl_config":          resourceSTLConfig(),
+			"hsdp_stl_custom_cert":     resourceSTLCustomCert(),
+			"hsdp_stl_sync":            resourceSTLSync(),
 		},
 		DataSourcesMap: map[string]*schema.Resource{
 			"hsdp_iam_introspect":              dataSourceIAMIntrospect(),
@@ -188,13 +199,16 @@ func Provider(build string) *schema.Provider {
 			"hsdp_iam_org":                     dataSourceIAMOrg(),
 			"hsdp_iam_proposition":             dataSourceIAMProposition(),
 			"hsdp_iam_application":             dataSourceIAMApplication(),
-			"hsdp_credentials_access":          dataSourceS3CredentialsAccess(),
-			"hsdp_credentials_policy":          dataSourceCredentialsPolicy(),
+			"hsdp_s3creds_access":              dataSourceS3CredsAccess(),
+			"hsdp_s3creds_policy":              dataSourceS3CredsPolicy(),
 			"hsdp_config":                      dataSourceConfig(),
 			"hsdp_container_host_subnet_types": dataSourceContainerHostSubnetTypes(),
 			"hsdp_cdr_fhir_store":              dataSourceCDRFHIRStore(),
+			"hsdp_pki_root":                    dataSourcePKIRoot(),
+			"hsdp_pki_policy":                  dataSourcePKIPolicy(),
+			"hsdp_stl_device":                  dataSourceSTLDevice(),
 		},
-		ConfigureContextFunc: providerConfigure,
+		ConfigureContextFunc: providerConfigure(build),
 	}
 }
 
@@ -206,7 +220,7 @@ func init() {
 		"environment":         "The HSDP environment to configure for",
 		"iam_url":             "The HSDP IAM instance URL",
 		"idm_url":             "The HSDP IDM instance URL",
-		"credentials_url":     "The HSDP S3 Credentials instance URL",
+		"s3creds_url":         "The HSDP S3 Credentials instance URL",
 		"oauth2_client_id":    "The OAuth2 client id",
 		"oauth2_password":     "The OAuth2 password",
 		"service_id":          "The service ID to use as Organization Admin",
@@ -230,49 +244,61 @@ func init() {
 	}
 }
 
-func providerConfigure(ctx context.Context, d *schema.ResourceData) (interface{}, diag.Diagnostics) {
-	// Warning or errors can be collected in a slice type
-	var diags diag.Diagnostics
+func providerConfigure(build string) schema.ConfigureContextFunc {
+	return func(ctx context.Context, d *schema.ResourceData) (interface{}, diag.Diagnostics) {
+		var diags diag.Diagnostics
 
-	config := &Config{}
+		config := &Config{}
 
-	config.Region = d.Get("region").(string)
-	config.Environment = d.Get("environment").(string)
-	config.IAMURL = d.Get("iam_url").(string)
-	config.IDMURL = d.Get("idm_url").(string)
-	config.OAuth2ClientID = d.Get("oauth2_client_id").(string)
-	config.OAuth2Secret = d.Get("oauth2_password").(string)
-	config.RootOrgID = d.Get("org_id").(string)
-	config.ServiceID = d.Get("service_id").(string)
-	config.ServicePrivateKey = d.Get("service_private_key").(string)
-	config.OrgAdminUsername = d.Get("org_admin_username").(string)
-	config.OrgAdminPassword = d.Get("org_admin_password").(string)
-	config.SharedKey = d.Get("shared_key").(string)
-	config.SecretKey = d.Get("secret_key").(string)
-	config.Debug = d.Get("debug").(bool)
-	config.DebugLog = d.Get("debug_log").(string)
-	config.S3CredsURL = d.Get("credentials_url").(string)
-	config.CartelHost = d.Get("cartel_host").(string)
-	config.CartelToken = d.Get("cartel_token").(string)
-	config.CartelSecret = d.Get("cartel_secret").(string)
-	config.CartelNoTLS = d.Get("cartel_no_tls").(bool)
-	config.CartelSkipVerify = d.Get("cartel_skip_verify").(bool)
-	config.RetryMax = d.Get("retry_max").(int)
-	config.UAAUsername = d.Get("uaa_username").(string)
-	config.UAAPassword = d.Get("uaa_password").(string)
-	config.UAAURL = d.Get("uaa_url").(string)
-	config.TimeZone = "UTC"
+		config.BuildVersion = build
+		config.Region = d.Get("region").(string)
+		config.Environment = d.Get("environment").(string)
+		config.IAMURL = d.Get("iam_url").(string)
+		config.IDMURL = d.Get("idm_url").(string)
+		config.OAuth2ClientID = d.Get("oauth2_client_id").(string)
+		config.OAuth2Secret = d.Get("oauth2_password").(string)
+		config.RootOrgID = d.Get("org_id").(string)
+		config.ServiceID = d.Get("service_id").(string)
+		config.ServicePrivateKey = d.Get("service_private_key").(string)
+		config.OrgAdminUsername = d.Get("org_admin_username").(string)
+		config.OrgAdminPassword = d.Get("org_admin_password").(string)
+		config.SharedKey = d.Get("shared_key").(string)
+		config.SecretKey = d.Get("secret_key").(string)
+		config.DebugLog = d.Get("debug_log").(string)
+		config.S3CredsURL = d.Get("s3creds_url").(string)
+		config.CartelHost = d.Get("cartel_host").(string)
+		config.CartelToken = d.Get("cartel_token").(string)
+		config.CartelSecret = d.Get("cartel_secret").(string)
+		config.CartelNoTLS = d.Get("cartel_no_tls").(bool)
+		config.CartelSkipVerify = d.Get("cartel_skip_verify").(bool)
+		config.RetryMax = d.Get("retry_max").(int)
+		config.UAAUsername = d.Get("uaa_username").(string)
+		config.UAAPassword = d.Get("uaa_password").(string)
+		config.UAAURL = d.Get("uaa_url").(string)
+		config.TimeZone = "UTC"
 
-	config.setupIAMClient()
-	config.setupS3CredsClient()
-	config.setupCartelClient()
-	config.setupConsoleClient()
+		config.setupIAMClient()
+		config.setupS3CredsClient()
+		config.setupCartelClient()
+		config.setupConsoleClient()
+		config.setupPKIClient()
+		config.setupSTLClient()
 
-	ma, err := jsonformat.NewMarshaller(false, "", "", jsonformat.STU3)
-	if err != nil {
-		return nil, diag.FromErr(err)
+		if config.DebugLog != "" {
+			debugFile, err := os.OpenFile(config.DebugLog, os.O_APPEND|os.O_CREATE|os.O_WRONLY, 0600)
+			if err != nil {
+				config.debugFile = nil
+			} else {
+				config.debugFile = debugFile
+			}
+		}
+
+		ma, err := jsonformat.NewMarshaller(false, "", "", jsonformat.STU3)
+		if err != nil {
+			return nil, diag.FromErr(err)
+		}
+		config.ma = ma
+
+		return config, diags
 	}
-	config.ma = ma
-
-	return config, diags
 }